+++ accounts/apps.py
@@ -29,9 +29,9 @@ class AppAccountsConfig(AppConfig):
         def connect_signals(sender, **kwargs):
             from django.db.models.signals import post_save
             from .models import CustomUser
-            from accounts.signals import set_admin_permissions
             
-            from accounts.signals import set_geodesy_permissions
+            from accounts.signals import set_admin_permissions
+            from accounts.signals import set_verifying_authority_permissions
 
             post_save.connect(set_admin_permissions, sender=CustomUser)
-            post_save.connect(set_geodesy_permissions, sender=CustomUser)
+            post_save.connect(set_verifying_authority_permissions, sender=CustomUser)