+++ accounts/signals.py
@@ -13,7 +13,11 @@ def set_admin_permissions(sender, instance, **kwargs):
 
     # ==> Add users to groups
     admin_users = ['kent.wheeler@landgate.wa.gov.au', 
-                'khandu.k@landgate.wa.gov.au']
+                    'khandu.k@landgate.wa.gov.au',
+                    'irek.baran@landgate.wa.gov.au',
+                    'brendon.hellmund@landgate.wa.gov.au',
+                    'vanessa.ung@landgate.wa.gov.au',
+                    'michael.kuhn@landgate.wa.gov.au']
     for email in admin_users:
         try:
             user_id = CustomUser.objects.get(email = email)
@@ -21,23 +25,23 @@ def set_admin_permissions(sender, instance, **kwargs):
         except ObjectDoesNotExist:
             pass
 
-# permissions to geodesy_group
+# permissions to verifying_authority_group - permission for TOTPDevice excluded
 @receiver(post_save, sender=CustomUser)
-def set_geodesy_permissions(sender, instance, **kwargs):
-    geodesy_group = Group.objects.filter(name='Geodesy').first()
+def set_verifying_authority_permissions(sender, instance, **kwargs):
+    verifying_authority_group = Group.objects.filter(name='Verifying_Authority').first()
     permissions = Permission.objects.exclude(codename__startswith='delete_')
-    # geodesy_group.permissions.add(*permissions)
     # ==> Add users to groups
-    geodesy_users = ['irek.baran@landgate.wa.gov.au',
+    landgate_users = ['kent.wheeler@landgate.wa.gov.au',
                     'tony.castelli@landgate.wa.gov.au',
+                    'david.martin@landgate.wa.gov.au',
+                    'martin.cole@landgate.wa.gov.au',
                     'brendon.hellmund@landgate.wa.gov.au',
-                    'vanessa.ung@landgate.wa.gov.au',
-                    'khandu.k@landgate.wa.gov.au',
-                    'michael.kuhn@landgate.wa.gov.au']
-    for email in geodesy_users:
+                    'khandu.k@landgate.wa.gov.au',]
+    
+    for email in landgate_users:
         try:
             user_id = CustomUser.objects.get(email = email)
-            geodesy_group.user_set.add(user_id)
+            verifying_authority_group.user_set.add(user_id)
         except ObjectDoesNotExist:
             pass
 
@@ -45,12 +49,4 @@ def set_geodesy_permissions(sender, instance, **kwargs):
     # Add all permissions except for the excluded model
     for permit in permissions:
         if permit.content_type != content_type_to_exclude or permit.codename.startswith('view'):
-            geodesy_group.permissions.add(permit)
-
-
-# # Get the content type for the MedjilTOTPDevice
-# content_type = ContentType.objects.get_for_model(MedjilTOTPDevice)
-# # allow only view permission for TOTP devices for geodesy_group
-# view_permission = Permission.objects.get(codename='view_medjiltotpdevice', content_type = content_type)
-# # Set permission
-# geodesy_group.permissions.set([view_permission])
+            verifying_authority_group.permissions.add(permit)